boring: convert time_t from i32 to i64
When build for 32bit ARMv7 with musl libc in the OHOS toolchain time_t
is i32 while it is expected to be i64. Converting it to i64 here.

warning: boring-sys@4.6.0: fetching boringssl git submodule
   Compiling boring v4.6.0
warning: use of deprecated type alias `libc::time_t`: This type is changed to 64-bit in musl 1.2.0, we'll follow that change in the future release. See #1848 for more info.
  --> boring/src/asn1.rs:29:35
   |
29 | use libc::{c_char, c_int, c_long, time_t};
   |                                   ^^^^^^
   |
   = note: `#[warn(deprecated)]` on by default

error[E0308]: mismatched types
     --> boring/src/asn1.rs:320:68
      |
320   |             let handle = cvt_p(ffi::ASN1_TIME_set(ptr::null_mut(), time))?;
      |                                ------------------                  ^^^^ expected `i64`, found `i32`
      |                                |
      |                                arguments to this function are incorrect
      |
note: function defined here
     --> /home/stefan/huawei/signal/boring/target/armv7-unknown-linux-ohos/release/build/boring-sys-f28cc373372f55e7/out/bindings.rs:10562:12
      |
10562 |     pub fn ASN1_TIME_set(s: *mut ASN1_TIME, time: time_t) -> *mut ASN1_TIME;
      |            ^^^^^^^^^^^^^
help: you can convert an `i32` to an `i64`

+++ boring/src/x509/verify.rs
@@ -170,6 +170,6 @@ impl X509VerifyParamRef {
     /// store_builder.param_mut().set_time(seconds_since_epoch);
     /// ```
     pub fn set_time(&mut self, unix_time: time_t) {
-        unsafe { ffi::X509_VERIFY_PARAM_set_time(self.as_ptr(), unix_time) }
+        unsafe { ffi::X509_VERIFY_PARAM_set_time(self.as_ptr(), unix_time as i64) }
     }
 }
